update some scripts in lib

+++ lib/copy-file-history-repo-to-home.sh
@@ -5,15 +5,17 @@ home_dst="$2"
 
 [ ! -e $repo_src ] && exit
 
-echo $repo_src -\> $home_dst
-mkdir -p $(dirname "$home_dst")
+if [ ! -e $home_dst ] || ! diff $repo_src $home_dst >/dev/null; then
+    echo $repo_src -\> $home_dst
+    mkdir -p $(dirname "$home_dst")
 
-(
-    cat $repo_src
-    if [ -e $home_dst ]; then
-      diff -u $repo_src $home_dst | grep -a -e '^+' | grep -a -v -e '^+++' | cut -b2-
-    fi
-) >| $home_dst.merged
+    (
+        cat $repo_src
+        if [ -e $home_dst ]; then
+          diff -u $repo_src $home_dst | grep -a -e '^+' | grep -a -v -e '^+++' | cut -b2-
+        fi
+    ) >| $home_dst.merged
 
-mv $home_dst.merged $home_dst
+    mv $home_dst.merged $home_dst
+fi
