+++ lib/copy-file-normal-home-to-repo.sh
@@ -5,8 +5,10 @@ repo_dst="$2"
 
 [ ! -e $home_src ] && exit
 
-echo $home_src -\> $repo_dst
-mkdir -p $(dirname "$repo_dst")
+if [ ! -e $repo_dst ] || ! diff $home_src $repo_dst >/dev/null; then
+    echo $home_src -\> $repo_dst
+    mkdir -p $(dirname "$repo_dst")
 
-cp $home_src $repo_dst
+    cp $home_src $repo_dst
+fi
