+++ lib/copy-file-normal-repo-to-home.sh
@@ -5,8 +5,10 @@ home_dst="$2"
 
 [ ! -e $repo_src ] && exit
 
-echo $repo_src -\> $home_dst
-mkdir -p $(dirname "$home_dst")
+if [ ! -e $home_dst ] || ! diff $repo_src $home_dst >/dev/null; then
+    echo $repo_src -\> $home_dst
+    mkdir -p $(dirname "$home_dst")
 
-cp $repo_src $home_dst
+    cp $repo_src $home_dst
+fi
